test the server-status endpoints for correctness and refactor to use one server router category & class for retrieving server provisioning information
