+++ minecraft-platform-backend-api/setup.cfg
@@ -40,6 +40,7 @@ install_requires =
     boto3
     uvicorn
     mypy_boto3_cloudformation
+    mypy_boto3_stepfunctions
 
 [options.packages.find]
 where = src
