+++ minecraft-platform-backend-api/src/minecraft_paas_api/schemas/server_destroy_post.py
@@ -1,0 +1,10 @@
+"""Moduel defines pydantic module for the destroy server post requests."""
+
+from pydantic import PositiveInt
+from pydantic.main import BaseModel
+
+
+class DestroyServer(BaseModel):
+    """Response model for the `/destroy-server-after-seconds` endpoint."""
+
+    time_to_wait_before_destroying_server: PositiveInt
